Refactored model operations

+++ app/Http/Controllers/TodosController.php
@@ -14,7 +14,7 @@ class TodosController extends Controller
      */
     public function index()
     {
-        return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
     }
 
     
@@ -31,23 +31,15 @@ class TodosController extends Controller
             'completed' => ['required', 'boolean'],
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed
+        ]);
 
         return response($todo, 201);
     }
-
-    /**
-     * Display the specified resource.
-     *
-     * @param  \App\Todo  $todo
-     * @return \Illuminate\Http\Response
-     */
-    public function show(Todo $todo)
-    {
-        //
-    }
-
-  
+ 
 
     /**
      * Update the specified resource in storage.
@@ -58,6 +50,10 @@ class TodosController extends Controller
      */
     public function update(Request $request, Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+        
         $data = $request->validate([
             'title' => ['required', 'string'],
             'completed' => ['required', 'boolean'],
@@ -76,6 +72,10 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
 
         return response('Deleted todo successfully', 200);
@@ -94,7 +94,7 @@ class TodosController extends Controller
             'completed' => ['required', 'boolean'],
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
         
         return response('Update successful', 200);
     }
@@ -108,10 +108,28 @@ class TodosController extends Controller
      */
     public function destroyCompleted(Request $request)
     { 
+
         $request->validate([
             'todos' => ['required', 'array']
         ]);
+
+        $todosToDelete = $request->todos;
+         
+        $userTodoIds = auth()->user()->todos->map(function($todo) {
+            return $todo->id;
+        });
+
+        // Check if all todos of these users contain the ids he or she wants to delete 
+        $valid = collect($todosToDelete)->every(function($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        }); 
+
+        if(!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+
         Todo::destroy($request->todos);
-        return response('Deleted', 200);
+
+        return response()->json('Deleted', 200);
     }
 }
